+++ app/controllers/users_controller.rb
@@ -1,0 +1,3 @@
+class UsersController < Devise::RegistrationsController
+  invisible_captcha only: :create
+end