+++ config/routes.rb
@@ -2,7 +2,9 @@ require "sidekiq/web"
 
 Rails.application.routes.draw do
   scope "(:locale)", locale: /#{I18n.available_locales.join("|")}/ do
-    devise_for :users
+    devise_for :users, controllers: {
+      registrations: "users"
+    }
 
     resource :about, only: :show, controller: :about
     resource :conduct, only: :show
@@ -85,6 +87,7 @@ Rails.application.routes.draw do
   end
 
   namespace :webhooks do
+    resource :postmark, only: :create, controller: :postmark
     resource :revenuecat, only: :create, controller: :revenue_cat
   end