+++ db/migrate/20220712221543_add_temporary_open_startup_metric_migration_column.rb
@@ -1,0 +1,5 @@
+class AddTemporaryOpenStartupMetricMigrationColumn < ActiveRecord::Migration[7.0]
+  def change
+    add_column :open_startup_metrics, :data_json, :jsonb
+  end
+end